ThirdRefactorTestCasesOptimized
