+++ src/test/java/selenium/dYanez/AutomationPractice.java
@@ -1,7 +1,6 @@
 package selenium.dYanez;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
-import org.hamcrest.core.StringContains;
 import org.junit.*;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
@@ -9,7 +8,6 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
@@ -26,6 +24,7 @@ public class AutomationPractice {
     @Before
     public void init(){
         driver = new ChromeDriver();
+        driver.get("http://automationpractice.com/");
         driver.manage().window().maximize();
         //wait implicit for list result visible cada vez que vaya a buscar al elemento
         driver.manage().timeouts().implicitlyWait(20, TimeUnit.SECONDS);
@@ -35,8 +34,7 @@ public class AutomationPractice {
     @Test
     public void atc01CargarPaginaPrincipal(){
         System.out.println("Test Case 1");
-        //Cargar la página
-        driver.get("http://automationpractice.com/");
+
         //validate string title pestaña of the page
         Assert.assertEquals("My Store",driver.getTitle());
 
@@ -59,8 +57,7 @@ public class AutomationPractice {
     @Test
     public void atc02CargarPaginaPrincipal2(){
         System.out.println("Test Case 2");
-        //cargar page
-        driver.get("http://automationpractice.com/");
+
         //validate string title pestaña of the page
         Assert.assertEquals("My Store",driver.getTitle());
         //element field write "printed chiffon dress"
@@ -88,8 +85,7 @@ public class AutomationPractice {
     @Test
     public void atc03MensajeProductoNoEncontrado() throws InterruptedException {
         System.out.println("Test Case 3");
-        //cargar page
-        driver.get("http://automationpractice.com/");
+
         //validate string title pestaña of the page
         Assert.assertEquals("My Store", driver.getTitle());
         //element field write "McBook"
@@ -109,16 +105,12 @@ public class AutomationPractice {
         //validate text msg
         Assert.assertEquals(titleText,"No results were found for your search \"McBook\"");
 
-
-
     }
-
     @Test
     public void atc04FindProductInListDymanic() {
         //En la caja de busqueda poner blo y seleccionar lo pre-buscado
         System.out.println("Test Case 4");
-        //cargar page
-        driver.get("http://automationpractice.com/");
+
         //Poner en caja de busqueda la palabra blo
         String wordSearchDress = "blo";
         WebElement searchFieldDress = driver.findElement(By.cssSelector("#search_query_top"));
@@ -137,8 +129,7 @@ public class AutomationPractice {
     public void atc05FindProductInListDymanic() {
         //En la caja de busqueda poner blo y seleccionar lo pre-buscado
         System.out.println("Test Case 5");
-        //cargar page
-        driver.get("http://automationpractice.com/");
+
         //Poner en caja de busqueda la palabra blo
         String wordSearchDress = "blo";
         WebElement searchFieldDress = driver.findElement(By.cssSelector("#search_query_top"));
@@ -169,49 +160,7 @@ public class AutomationPractice {
 
         Assert.assertEquals("Product successfully added to your shopping cart", productCart.getText());
     }
-    @Test
-    public void atc06AgregarReviewFrankPage() throws InterruptedException {
-        //En la caja de busqueda poner blo y seleccionar lo pre-buscado
-        System.out.println("Test Case 6");
-        //cargar page
-        driver.get("http://automation.frankluzon.com/");
 
-        WebElement productClick = driver.findElement(By.xpath("//*[@id=\"post-27\"]/div/div[3]/ul/li[1]/a/div[1]/img"));
-        productClick.click();
-
-
-        WebElement reviewClick = driver.findElement(By.xpath("//*[@id=\"tab-title-reviews\"]/a"));
-        reviewClick.click();
-
-        WebElement startRatingClick = driver.findElement(By.xpath("//*[@id=\"commentform\"]/div/p/span/a[4]"));
-        startRatingClick.click();
-
-
-        WebElement writeComment = driver.findElement(By.xpath("//*[@id=\"comment\"]"));
-        writeComment.sendKeys("lorem ipsum lorem ipsum lorem ipsum");
-
-        WebElement writeName = driver.findElement(By.xpath("//input[@id=\"author\"]"));
-        writeName.sendKeys("DiegoY");
-
-        WebElement writeEmail = driver.findElement(By.xpath("//input[@id='email']"));
-        writeEmail.sendKeys("diegotsoft@gmail.cl");
-
-        WebElement checkboxClick = driver.findElement(By.xpath("//*[@id=\"wp-comment-cookies-consent\"]"));
-        checkboxClick.click();
-
-        WebElement submitClick = driver.findElement(By.xpath("//input[@id='submit']"));
-        submitClick.click();
-
-
-        String textResult = driver.findElement( By.xpath("//div[@class=\"wp-die-message\"]")).getText();
-        System.out.println(textResult);
-
-
-        //validate msg
-        String msgExpected = "Your review is awaiting approval";
-        Assert.assertEquals(msgExpected,textResult);
-
-    }
 
 
     @After
